+++ dkplayer-sample/src/main/java/xyz/doikki/dkplayer/widget/component/MyDanmakuView.java
@@ -150,7 +150,7 @@ public class MyDanmakuView extends DanmakuView implements ControlComponent {
     }
 
     @Override
-    public void onProgressChanged(int duration, int position) {
+    public void onProgressChanged(long duration, long position) {
 
     }
